refactor(IT-Cotato#248): 중고거래 스크랩 컨트롤러, 서비스 분리
