+++ src/main/java/com/cotato/kampus/domain/product/application/ProductService.java
@@ -52,7 +52,6 @@ public class ProductService {
 	private final ProductCategoryMappingManager productCategoryMappingManager;
 	private final ProductFinder productFinder;
 	private final ProductScrapFinder productScrapFinder;
-	private final ProductScrapManager productScrapManager;
 	private final ProductPhotoFinder productPhotoFinder;
 	private final ProductCategoryMappingFinder productCategoryMappingFinder;
 	private final ProductDtoMapper productDtoMapper;
@@ -144,51 +143,6 @@ public class ProductService {
 		productManager.update(product);
 		productCategoryMappingManager.saveAll(productId, categoryIds);
 		productPhotoManager.appendAll(product.getId(), imageUrls);
-	}
-
-
-	@Transactional
-	public void addScrap(Long productId) {
-		// 1. 유저 조회/검증
-		UserDto user = apiUserResolver.getCurrentUserDto();
-		userValidator.validateStudentVerification(user);
-
-		// 2. 상품 조회/검증
-		Product product = productFinder.findById(productId);
-		product.validateNotDeleted();
-
-		// 3. 스크랩 여부 검증
-		boolean isScrapped = productScrapFinder.isScrapped(productId, user.id());
-		if(isScrapped) {
-			throw new AppException(ErrorCode.ALREADY_SCRAPPED_PRODUCT);
-		}
-
-		// 4. 스크랩 추가
-		productScrapManager.append(productId, user.id());
-
-		// 5. 상품 스크랩 수 반영
-		Product scrappedProduct = product.increaseScrapCount();
-		productManager.update(scrappedProduct);
-	}
-
-	@Transactional
-	public void removeScrap(Long productId) {
-		// 1. 유저, 상품 조회
-		UserDto user = apiUserResolver.getCurrentUserDto();
-		Product product = productFinder.findById(productId);
-
-		// 2. 스크랩 여부 검증
-		boolean isScrapped = productScrapFinder.isScrapped(productId, user.id());
-		if(!isScrapped) {
-			throw new AppException(ErrorCode.PRODUCT_SCRAP_NOT_FOUND);
-		}
-
-		// 3. 스크랩 취소
-		productScrapManager.delete(product.getId(), user.id());
-
-		// 4. 상품 스크랩 수 감소
-		Product unscrappedProduct = product.decreaseScrapCount();
-		productManager.update(unscrappedProduct);
 	}
 
 	@Transactional
